Настройка сборки и разбиение кода на компоненты
